facturas

+++ pasteleria/src/negocio/Facturas/SAFacturaImp.java
@@ -31,8 +31,7 @@ public class SAFacturaImp implements SAFactura {
 		DAOLineaFactura daoLineaFactura = FactoriaAbstractaIntegracion.getInstance().crearDAOLineaFactura();
 		DAOCliente daoCliente = FactoriaAbstractaIntegracion.getInstance().crearDAOCliente();
 		DAOProducto daoProducto = FactoriaAbstractaIntegracion.getInstance().crearDAOProducto();
-		// DAOEmpleado daoEmpleado =
-		// FactoriaAbstractaIntegracion.getInstance().crearDAOEmpleado();
+		DAOEmpleado daoEmpleado = FactoriaAbstractaIntegracion.getInstance().crearDAOEmpleado();
 
 		TFactura factura;
 		ArrayList<TLineaFactura> lineas_factura_definitivas = new ArrayList<TLineaFactura>();
@@ -40,42 +39,46 @@ public class SAFacturaImp implements SAFactura {
 		int precio_total = 0;
 		int id_lineaFactura = 0;
 		int id_factura = -1;
-		TCliente cliente = daoCliente.buscarCliente(datos.getIdCliente());
-		// TEmpleado vendedor = daoEmpleado.buscarEmpleado(datos.getIdVendedor());
-		boolean exito = true;
-		// if (cliente != null && vendedor != null*) {
-		int i = 0;
-		// Recorremos las lineas de factura de nuestra nueva factura
-		ArrayList<TLineaFactura> lineas_factura_por_comprobar = datos.getProductos();
-		while (i < lineas_factura_por_comprobar.size() && exito) {
-			TLineaFactura linea_factura = lineas_factura_por_comprobar.get(i);
-			// comprobamos el producto
-			TProducto producto = daoProducto.buscarProducto(linea_factura.getIdProducto());
-			if (producto == null || linea_factura.getCantidad() < 0 || producto.getStock() < 0
-					|| producto.getPrecio() < 0) {
-				exito = false;
-			} else {
-				if (producto.getStock() < linea_factura.getCantidad()) {
-					linea_factura.setCantidadProducto(producto.getStock());
-				}
-				id_lineaFactura = linea_factura.getIdLinea();
-				precio_total += producto.getPrecio() * linea_factura.getCantidad();
-				producto.setStock(producto.getStock() - linea_factura.getCantidad());
-				// guardamos la linea de factura
-				lineas_factura_definitivas.add(linea_factura);
-			}
-			i++;
-		}
 
-		// }
+		TCliente cliente = daoCliente.buscarCliente(datos.getIdCliente());
+		TEmpleado vendedor = daoEmpleado.buscarEmpleado(datos.getIdVendedor());
+
+		boolean exito = true;
+		if (cliente != null /* && vendedor != null */) {
+			int i = 0;
+			// Recorremos las lineas de factura de nuestra nueva factura
+			ArrayList<TLineaFactura> lineas_factura_por_comprobar = datos.getProductos();
+			while (i < lineas_factura_por_comprobar.size() && exito) {
+				TLineaFactura linea_factura = lineas_factura_por_comprobar.get(i);
+				// comprobamos el producto
+				TProducto producto = daoProducto.buscarProducto(linea_factura.getIdProducto());
+				if (producto == null || linea_factura.getCantidad() < 0 || producto.getStock() < 0
+						|| producto.getPrecio() < 0) {
+					exito = false;
+				} else {
+					if (producto.getStock() < linea_factura.getCantidad()) {
+						linea_factura.setCantidadProducto(producto.getStock());
+					}
+					id_lineaFactura = linea_factura.getIdLinea();
+					precio_total += producto.getPrecio() * linea_factura.getCantidad();
+					producto.setStock(producto.getStock() - linea_factura.getCantidad());
+					daoProducto.actualizarProducto(producto);
+					// guardamos la linea de factura
+					lineas_factura_definitivas.add(linea_factura);
+				}
+				i++;
+			}
+
+		}
 		if (lineas_factura_definitivas.size() > 0 && exito) {// si hay lineas válidas y no hay problemas en la factura
 			id_factura = daoFactura.crearFactura(new TFactura(0, precio_total, datos, true));
 			for (TLineaFactura lf : lineas_factura_definitivas) {
 				lf.setIdFactura(id_factura);
 				daoLineaFactura.crearLineaFactura(lf);
+				int uuuu=0;
 			}
-//			vendedor.setNumVentas(vendedor.getNumVentas() + 1);
-//			vendedor.update();
+			// vendedor.setNumVentas(vendedor.getNumVentas() + 1);
+			// daoEmpleado.actualizarEmpleado(vendedor);
 		}
 		return id_factura;
 	}
@@ -83,7 +86,7 @@ public class SAFacturaImp implements SAFactura {
 	public boolean anadirProducto(TLineaFactura linea, Carrito c) {
 		DAOProducto daoProducto = FactoriaAbstractaIntegracion.getInstance().crearDAOProducto();
 		TProducto producto = daoProducto.buscarProducto(linea.getIdProducto());
-		if (producto != null) {
+		if (producto != null && producto.getActivo()) {
 			c.anadirProducto(linea);
 		} else {
 			return false;
@@ -95,7 +98,7 @@ public class SAFacturaImp implements SAFactura {
 		boolean exito = true;
 		DAOProducto daoProducto = FactoriaAbstractaIntegracion.getInstance().crearDAOProducto();
 		TProducto producto = daoProducto.buscarProducto(linea.getIdProducto());
-		if (producto != null) {
+		if (producto != null && producto.getActivo()) {
 			exito = c.eliminarProducto(linea);
 		} else {
 			exito = false;
